Web/web05/requirements.txt
Flask==2.0.3
gunicorn==20.1.0
requests==2.28.1
validators==0.20.0
tldextract==3.4.0
dnspython==2.2.0
